requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
black==18.6b4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cleo==0.6.5
click==6.7
cryptography==2.2.2
idna==2.7
masonite-cli==2.0.9
more-itertools==4.2.0
pastel==0.1.0
pluggy==0.6.0
py==1.5.4
pycparser==2.18
pylev==1.3.0
pytest==3.6.2
requests==2.19.1
six==1.11.0
toml==0.9.4
urllib3==1.23
